MiniRAG/requirements.txt
accelerate
aiofiles
aiohttp
configparser
graspologic
json_repair
httpx

# database packages
networkx
nltk

# Basic modules
numpy
pipmaster
pydantic

# File manipulation libraries
PyPDF2
python-docx
python-dotenv
python-pptx
rouge

setuptools
tenacity


# LLM packages (openai>=1.56 works with httpx 0.28+; older openai + httpx>=0.28 causes
# TypeError: unexpected keyword argument 'proxies')
openai>=1.56.0
httpx>=0.27.0
tiktoken
tqdm
xxhash

# Extra libraries are installed when needed using pipmaster
